+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - isequal, isequaln, isequalto extended to manage string array.
     - transpose, ctranspose extended to manage string array.
     - horzcat, vertcat extended to manage string array.
+    - double extended to manage string array.
 
   * internal API C++ methods renamed:
     - "stringConstructor" --> "characterArrayConstructor".